Replace Arial Font With Lato
It's unclear to me whether distribution of arial was strictly permissible; it may depend on the source of the file but where the file came from originally is also unclear; removing the font and replacing it with Lato sourced from https://fonts.google.com/specimen/Lato and available under the SIL Open Font License.

Compared to the other fonts I sampled lato looked very clear at a lot of sizes and also had a small file size.